rename shad dashboard to admin
