--- src/app/dashboard/page.tsx
@@ -1,27 +1,0 @@
-import { AppSidebar } from "~/components/app-sidebar"
-import { ChartAreaInteractive } from "~/components/chart-area-interactive"
-import { DataTable } from "~/components/data-table"
-import { SectionCards } from "~/components/section-cards"
-import { SiteHeader } from "~/components/site-header"
-import {
-  SidebarInset,
-  SidebarProvider,
-} from "~/components/ui/sidebar"
-
-import data from "./data.json"
-
-export default function Page() {
-  return (
-    <>
-        <div className="flex flex-1 flex-col pt-15">
-          <div className="@container/main flex flex-1 flex-col gap-2">
-            <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
-              <div className="px-4 lg:px-6">
-              </div>
-              <DataTable data={data} />
-            </div>
-          </div>
-        </div>
-    </>
-  )
-}
